Remove media sub-action files and modify log
Removed UploadMediaSubAction files from both Category and Product sub-folders under the SubActions directory. These files were redundant and not needed due to the restructuring of media upload functionality to the central Media library. Also, a new error is logged in the Laravel log file, concerning a temporary failure in resolving the host name of the database, which seems to indicate a temporary network issue or configuration problem.